requirements_db.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geoalchemy2==0.14.2
shapely==2.0.2
